Remove incorrect detection of simple fmt cases
This was intended to detect the simple cases "{var}" and "{var:?}" and emit them
as straightforward Display::fmt and Debug::fmt method calls rather than a write
macro, but this was only to keep the generated code simple and is not important.
It was broken by the changes to how we parse the fmt attr in 1.0.7.

Fixes #53.