Prefer core crate in macro expansions
Prefer core crate to std crate when expanding macros.  Like before,
the main motivation is easier support of no_std.

Two symbols needed to be handled specially.  Firstly, re-export
std::error::Error from thiserror::__private so that macro expansion
can use it without having to refer to std crate.  The idea here is
that in future commits, based on selected features, either std::error
or core::error will be used.

Secondly, leave std::backtrace::Backtrace as is since it doesn’t have
no_std equivalent.  Thankfully, the type is only used when user type
refers to it so it should not be a hindrance for no_std support.

While at it, use leading :: for core and std paths so they don’t
conflict with modules of the same name.