Add ability to capture `SpanTrace`s
This was inspired by reading a blog post that showed some shortcomings
of thiserror compared to snafu in capturing backtraces.

The approach here captures a SpanTrace created by the tracing-error
crate. https://docs.rs/tracing-error/

Note: this is a fairly mechanical copy of the relevant backtrace methods
for discussion. I have not yet checked whether theres simplifications
that make sense. It also is problematic due to the unstable version of
the tracing crates, so it's doubtful that this should be merged as-is.

My goal with this PR is to start a conversation about how it might work.

Fixes #400